requirements.txt
ujson==5.10.0
pytz==2024.1
termcolor==2.4.0
python-dotenv==1.0.1
aiogram==3.4.1
openai==1.40.3
pydantic==2.5.3
redis==5.0.8
redisvl==0.3.0
httpx==0.27.0
apscheduler==3.10.4